\name{sparseGFL-package}
\alias{sparseGFL-package}
\alias{sparseGFL}
\docType{package}
\title{
\packageTitle{sparseGFL}
}
\description{
\packageDescription{sparseGFL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sparseGFL}
\packageIndices{sparseGFL}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{sparseGFL}

Maintainer: \packageMaintainer{sparseGFL}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{regression, optimize}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
