\name{sparseGFL-package}
\alias{sparseGFL-package}
\alias{sparseGFL}
\docType{package}
\title{
\packageTitle{sparseGFL}
}
\description{
\packageDescription{sparseGFL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sparseGFL}
\packageIndices{sparseGFL}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{sparseGFL}

Maintainer: \packageMaintainer{sparseGFL}
}
\references{
Degras (2020). Sparse group fused lasso for model segmentation: a hybrid approach. \emph{Advances in Data Analysis and Classification.} \url{https://doi.org/10.1007/s11634-020-00424-5}
}
\keyword{regression, optimize}
\seealso{
\code{\link{SGFL}}, \code{\link{SGFL.AXY}}, \code{\link{SGFL.SVAR}} 
}
\examples{
~~ simple examples of the most important functions ~~
}
